move time picker state components under time package
