+++ app/src/main/kotlin/edu/stanford/bdh/engagehf/health/time/TimePickerStateMapper.kt
@@ -1,4 +1,4 @@
-package edu.stanford.bdh.engagehf.health.bloodpressure.bottomsheet
+package edu.stanford.bdh.engagehf.health.time
 
 import edu.stanford.spezi.core.utils.DateFormat
 import edu.stanford.spezi.core.utils.DateFormatter
