+++ app/src/test/kotlin/edu/stanford/bdh/engagehf/health/time/TimePickerStateMapperTest.kt
@@ -1,4 +1,4 @@
-package edu.stanford.bdh.engagehf.health.bloodpressure.bottomsheet
+package edu.stanford.bdh.engagehf.health.time
 
 import com.google.common.truth.Truth.assertThat
 import edu.stanford.spezi.core.utils.DateFormat
